backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
paho-mqtt
pydantic
python-dotenv
cryptography
